requirements.txt
Flask
gunicorn
python-dotenv
requests
pymongo
openai
tenacity
flask
logging
datetime
requests
mysql-connector-python